refactor: simplify navigation hooks name
